AndroidGCS: Convert UAVObject IDs to longs from ints because they are meant to be unsigned uint32 which is unsupported by java.

+++ androidgcs/src/org/openpilot/uavtalk/UAVMetaObject.java
@@ -7,7 +7,7 @@ import java.util.ListIterator;
 
 public class UAVMetaObject extends UAVObject {
 
-	public UAVMetaObject(int objID, String name, UAVDataObject parent) throws Exception {
+	public UAVMetaObject(long objID, String name, UAVDataObject parent) throws Exception {
 		super(objID, true, name);
 		this.parent = parent;
