+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/ActuatorSettings.java
@@ -242,7 +242,7 @@ public class ActuatorSettings extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			ActuatorSettings obj = new ActuatorSettings();
@@ -256,13 +256,13 @@ public class ActuatorSettings extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public ActuatorSettings GetInstance(UAVObjectManager objMngr, int instID)
+	public ActuatorSettings GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (ActuatorSettings)(objMngr.getObject(ActuatorSettings.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x7D555646;
+	protected static final long OBJID = 0x7D555646;
 	protected static final String NAME = "ActuatorSettings";
 	protected static String DESCRIPTION = "Settings for the @ref ActuatorModule that controls the channel assignments for the mixer based on AircraftType";
 	protected static final boolean ISSINGLEINST = 1 == 1;
