+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/FlightPlanControl.java
@@ -112,7 +112,7 @@ public class FlightPlanControl extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			FlightPlanControl obj = new FlightPlanControl();
@@ -126,13 +126,13 @@ public class FlightPlanControl extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public FlightPlanControl GetInstance(UAVObjectManager objMngr, int instID)
+	public FlightPlanControl GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (FlightPlanControl)(objMngr.getObject(FlightPlanControl.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x53E3F180;
+	protected static final long OBJID = 0x53E3F180;
 	protected static final String NAME = "FlightPlanControl";
 	protected static String DESCRIPTION = "Control the flight plan script";
 	protected static final boolean ISSINGLEINST = 1 == 1;
