+++ androidgcs/src/org/openpilot/uavtalk/uavobjects/MagBias.java
@@ -115,7 +115,7 @@ public class MagBias extends UAVDataObject {
 	 * Do not use this function directly to create new instances, the
 	 * UAVObjectManager should be used instead.
 	 */
-	public UAVDataObject clone(int instID) {
+	public UAVDataObject clone(long instID) {
 		// TODO: Need to get specific instance to clone
 		try {
 			MagBias obj = new MagBias();
@@ -129,13 +129,13 @@ public class MagBias extends UAVDataObject {
 	/**
 	 * Static function to retrieve an instance of the object.
 	 */
-	public MagBias GetInstance(UAVObjectManager objMngr, int instID)
+	public MagBias GetInstance(UAVObjectManager objMngr, long instID)
 	{
 	    return (MagBias)(objMngr.getObject(MagBias.OBJID, instID));
 	}
 
 	// Constants
-	protected static final int OBJID = 0x5043E510;
+	protected static final long OBJID = 0x5043E510;
 	protected static final String NAME = "MagBias";
 	protected static String DESCRIPTION = "The gyro data.";
 	protected static final boolean ISSINGLEINST = 1 == 1;
